feat(cellgrid): implement onclick action in cellgrid #40
- move current onclick action in grid to cellgrid component
